+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ### general
 
+- strongly speed-up contingency table computation for heterogeneous=true and max_k=0/1
 - use the more compiler-friendly `stack` (introduced in Julia v1.9) instead of `hcat` for large numbers of columns (if available)
 - improve univariate pvalue filtering
 - remove performance bottleneck in three-way `adjust_df`